Separate the sitemap spider from the pages spider
